# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavez Prod Backend",
    "description": "Rest Api for Wavez Backend",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    },
    {
      "url": "https://wavez-dev.cf:3000/api"
    },
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/reviews/updateReviewsSequence": {
      "put": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Updates the sequence of the Reviews (Authorization: [ADMIN])",
        "operationId": "updateReviewSequence",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "sequence": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "description": "Updates the sequence of the reviews (use -1 as sequence to delete a featured review) [Authorization: ADMIN]",
        "responses": {
          "200": {
            "description": "updated sequence",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/messages/save": {
      "post": {
        "tags": [
          "Message"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Save a new message",
        "operationId": "saveMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "receiverId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "description": "Save a new message",
        "responses": {
          "200": {
            "description": "Saved message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "conversationId": {
                      "type": "string"
                    },
                    "sender": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}